requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-Mail==0.9.1
Flask-JWT-Extended==4.5.3
WTForms==3.0.1
psycopg2-binary==2.9.7
redis==5.0.1
celery==5.3.4
requests==2.31.0
python-dotenv==1.0.0
stripe==6.6.0
Pillow==10.0.1
email-validator==2.0.0
itsdangerous==2.1.2
google-genai==0.3.0
boto3==1.34.0
python-instagram-basic-display==1.3.7
facebook-sdk==3.1.0
gunicorn==21.2.0